Refactor codeActionProvider.ts
