+++ Extension/src/LanguageServer/client.ts
@@ -17,6 +17,7 @@ import { DocumentSymbolProvider } from './Providers/documentSymbolProvider';
 import { WorkspaceSymbolProvider } from './Providers/workspaceSymbolProvider';
 import { RenameProvider } from './Providers/renameProvider';
 import { FindAllReferencesProvider } from './Providers/findAllReferencesProvider';
+import { CodeActionProvider } from './Providers/codeActionProvider';
 // End provider imports
 
 import { LanguageClient, LanguageClientOptions, ServerOptions, NotificationType, TextDocumentIdentifier, RequestType, ErrorAction, CloseAction, DidOpenTextDocumentParams, Range, Position, DocumentFilter } from 'vscode-languageclient';
@@ -65,7 +66,7 @@ let warningChannel: vscode.OutputChannel;
 let diagnosticsCollectionIntelliSense: vscode.DiagnosticCollection;
 let diagnosticsCollectionCodeAnalysis: vscode.DiagnosticCollection;
 
-interface CodeActionDiagnosticInfo {
+export interface CodeActionDiagnosticInfo {
     // range/code are used to work around https://github.com/microsoft/vscode/issues/126393.
     // If that bug were fixed, then we could use the vscode.CodeAction.diagnostic directly.
     range: vscode.Range;
@@ -80,7 +81,7 @@ interface CodeActionPerUriInfo {
     workspaceEdits?: vscode.WorkspaceEdit[];
 }
 
-interface CodeActionCodeInfo {
+export interface CodeActionCodeInfo {
     uriToInfo: Map<string, CodeActionPerUriInfo>;
     fixAllTypeCodeAction?: vscode.CodeAction;
     disableAllTypeCodeAction?: vscode.CodeAction;
@@ -93,9 +94,9 @@ interface CodeActionAllInfo {
     removeAllCodeAction: vscode.CodeAction;
 }
 
-const codeAnalysisFileToCodeActions: Map<string, CodeActionDiagnosticInfo[]> = new Map<string, CodeActionDiagnosticInfo[]>();
-const codeAnalysisCodeToFixes: Map<string, CodeActionCodeInfo> = new Map<string, CodeActionCodeInfo>();
-const codeAnalysisAllFixes: CodeActionAllInfo = {
+export const codeAnalysisFileToCodeActions: Map<string, CodeActionDiagnosticInfo[]> = new Map<string, CodeActionDiagnosticInfo[]>();
+export const codeAnalysisCodeToFixes: Map<string, CodeActionCodeInfo> = new Map<string, CodeActionCodeInfo>();
+export const codeAnalysisAllFixes: CodeActionAllInfo = {
     fixAllCodeAction: {
         title: localize("fix_all_code_analysis_problems", "Fix all code analysis problems"),
         command: { title: 'RemoveCodeAnalysisProblems', command: 'C_Cpp.RemoveCodeAnalysisProblems',
@@ -598,19 +599,6 @@ interface PublishDiagnosticsParams {
     diagnostics: Diagnostic[];
 }
 
-interface GetCodeActionsRequestParams {
-    uri: string;
-    range: Range;
-}
-
-interface CodeActionCommand {
-    localizeStringParams: LocalizeStringParams;
-    command: string;
-    arguments?: any[];
-    edit?: TextEdit;
-    uri?: string;
-}
-
 interface ShowMessageWindowParams {
     type: number;
     localizeStringParams: LocalizeStringParams;
@@ -806,7 +794,6 @@ const QueryCompilerDefaultsRequest: RequestType<QueryCompilerDefaultsParams, con
 const QueryTranslationUnitSourceRequest: RequestType<QueryTranslationUnitSourceParams, QueryTranslationUnitSourceResult, void, void> = new RequestType<QueryTranslationUnitSourceParams, QueryTranslationUnitSourceResult, void, void>('cpptools/queryTranslationUnitSource');
 const SwitchHeaderSourceRequest: RequestType<SwitchHeaderSourceParams, string, void, void> = new RequestType<SwitchHeaderSourceParams, string, void, void>('cpptools/didSwitchHeaderSource');
 const GetDiagnosticsRequest: RequestType<void, GetDiagnosticsResult, void, void> = new RequestType<void, GetDiagnosticsResult, void, void>('cpptools/getDiagnostics');
-const GetCodeActionsRequest: RequestType<GetCodeActionsRequestParams, CodeActionCommand[], void, void> = new RequestType<GetCodeActionsRequestParams, CodeActionCommand[], void, void>('cpptools/getCodeActions');
 export const GetDocumentSymbolRequest: RequestType<GetDocumentSymbolRequestParams, LocalizeDocumentSymbol[], void, void> = new RequestType<GetDocumentSymbolRequestParams, LocalizeDocumentSymbol[], void, void>('cpptools/getDocumentSymbols');
 export const GetSymbolInfoRequest: RequestType<WorkspaceSymbolParams, LocalizeSymbolInformation[], void, void> = new RequestType<WorkspaceSymbolParams, LocalizeSymbolInformation[], void, void>('cpptools/getWorkspaceSymbols');
 export const GetFoldingRangesRequest: RequestType<GetFoldingRangesParams, GetFoldingRangesResult, void, void> = new RequestType<GetFoldingRangesParams, GetFoldingRangesResult, void, void>('cpptools/getFoldingRanges');
@@ -1206,136 +1193,6 @@ export class DefaultClient implements Client {
                     this.innerLanguageClient = languageClient;
                     telemetry.logLanguageServerEvent("NonDefaultInitialCppSettings", this.settingsTracker.getUserModifiedSettings());
                     failureMessageShown = false;
-
-                    class CodeActionProvider implements vscode.CodeActionProvider {
-                        private client: DefaultClient;
-                        constructor(client: DefaultClient) {
-                            this.client = client;
-                        }
-
-                        public async provideCodeActions(document: vscode.TextDocument, range: vscode.Range | vscode.Selection, context: vscode.CodeActionContext, token: vscode.CancellationToken): Promise<(vscode.Command | vscode.CodeAction)[]> {
-                            return this.client.requestWhenReady(async () => {
-                                let r: Range;
-                                if (range instanceof vscode.Selection) {
-                                    if (range.active.isBefore(range.anchor)) {
-                                        r = Range.create(Position.create(range.active.line, range.active.character), Position.create(range.anchor.line, range.anchor.character));
-                                    } else {
-                                        r = Range.create(Position.create(range.anchor.line, range.anchor.character), Position.create(range.active.line, range.active.character));
-                                    }
-                                } else {
-                                    r = Range.create(Position.create(range.start.line, range.start.character), Position.create(range.end.line, range.end.character));
-                                }
-
-                                const params: GetCodeActionsRequestParams = {
-                                    range: r,
-                                    uri: document.uri.toString()
-                                };
-
-                                const commands: CodeActionCommand[] = await this.client.languageClient.sendRequest(GetCodeActionsRequest, params, token);
-                                const resultCodeActions: vscode.CodeAction[] = [];
-
-                                // Convert to vscode.CodeAction array
-                                commands.forEach((command) => {
-                                    const title: string = util.getLocalizedString(command.localizeStringParams);
-                                    let wsEdit: vscode.WorkspaceEdit | undefined;
-                                    let codeActionKind: vscode.CodeActionKind = vscode.CodeActionKind.QuickFix;
-                                    if (command.edit) {
-                                        codeActionKind = vscode.CodeActionKind.RefactorInline;
-                                        wsEdit = new vscode.WorkspaceEdit();
-                                        wsEdit.replace(document.uri, vscodeRange(command.edit.range), command.edit.newText);
-                                    } else if (command.command === "C_Cpp.RemoveAllCodeAnalysisProblems" && command.uri !== undefined) {
-                                        const vsCodeRange: vscode.Range = vscodeRange(r);
-                                        const codeActions: CodeActionDiagnosticInfo[] | undefined = codeAnalysisFileToCodeActions.get(command.uri);
-                                        if (codeActions === undefined) {
-                                            return;
-                                        }
-                                        const fixCodeActions: vscode.CodeAction[] = [];
-                                        const disableCodeActions: vscode.CodeAction[] = [];
-                                        const removeCodeActions: vscode.CodeAction[] = [];
-                                        const docCodeActions: vscode.CodeAction[] = [];
-                                        const showClear: string = new CppSettings().clangTidyCodeActionShowClear;
-                                        for (const codeAction of codeActions) {
-                                            if (!codeAction.range.contains(vsCodeRange)) {
-                                                continue;
-                                            }
-                                            let codeActionCodeInfo: CodeActionCodeInfo | undefined;
-                                            if (codeAnalysisCodeToFixes.has(codeAction.code)) {
-                                                codeActionCodeInfo = codeAnalysisCodeToFixes.get(codeAction.code);
-                                            }
-                                            if (codeAction.fixCodeAction !== undefined) {
-                                                codeAction.fixCodeAction.isPreferred = true;
-                                                fixCodeActions.push(codeAction.fixCodeAction);
-                                                if (codeActionCodeInfo !== undefined) {
-                                                    if (codeActionCodeInfo.fixAllTypeCodeAction !== undefined &&
-                                                        (codeActionCodeInfo.uriToInfo.size > 1 ||
-                                                        codeActionCodeInfo.uriToInfo.values().next().value.workspaceEdits?.length > 1)) {
-                                                        fixCodeActions.push(codeActionCodeInfo.fixAllTypeCodeAction);
-                                                    }
-                                                }
-                                            }
-                                            let removeAllTypeAvailable: boolean = false;
-                                            if (codeActionCodeInfo !== undefined) {
-                                                if (codeActionCodeInfo.disableAllTypeCodeAction !== undefined) {
-                                                    disableCodeActions.push(codeActionCodeInfo.disableAllTypeCodeAction);
-                                                }
-                                                if (codeActionCodeInfo.removeAllTypeCodeAction !== undefined &&
-                                                    (codeActionCodeInfo.uriToInfo.size > 1 ||
-                                                    codeActionCodeInfo.uriToInfo.values().next().value.identifiers.length > 1)) {
-                                                    removeAllTypeAvailable = true;
-                                                }
-                                            }
-                                            if (showClear !== "None") {
-                                                if (!removeAllTypeAvailable || showClear === "AllAndAllTypeAndThis") {
-                                                    removeCodeActions.push(codeAction.removeCodeAction);
-                                                }
-                                                if (removeAllTypeAvailable && codeActionCodeInfo?.removeAllTypeCodeAction) {
-                                                    removeCodeActions.push(codeActionCodeInfo.removeAllTypeCodeAction);
-                                                }
-                                            }
-
-                                            if (codeActionCodeInfo === undefined || codeActionCodeInfo.docCodeAction === undefined) {
-                                                continue;
-                                            }
-                                            docCodeActions.push(codeActionCodeInfo.docCodeAction);
-                                        }
-                                        if (fixCodeActions.length > 0) {
-                                            resultCodeActions.push(...fixCodeActions);
-                                            if (codeAnalysisAllFixes.fixAllCodeAction.edit !== undefined) {
-                                                resultCodeActions.push(codeAnalysisAllFixes.fixAllCodeAction);
-                                            }
-                                        }
-                                        if (showClear !== "None") {
-                                            let showClearAllAvailable: boolean = false;
-                                            if ((codeActions.length > 1 || codeAnalysisFileToCodeActions.size > 1)) {
-                                                showClearAllAvailable = true;
-                                            }
-                                            if (!showClearAllAvailable || showClear !== "AllOnly") {
-                                                resultCodeActions.push(...removeCodeActions);
-                                            }
-                                            if (showClearAllAvailable) {
-                                                resultCodeActions.push(codeAnalysisAllFixes.removeAllCodeAction);
-                                            }
-                                        }
-                                        resultCodeActions.push(...disableCodeActions);
-                                        resultCodeActions.push(...docCodeActions);
-                                        return;
-                                    }
-                                    const vscodeCodeAction: vscode.CodeAction = {
-                                        title: title,
-                                        command: command.command === "edit" ? undefined : {
-                                            title: title,
-                                            command: command.command,
-                                            arguments: command.arguments
-                                        },
-                                        edit: wsEdit,
-                                        kind: codeActionKind
-                                    };
-                                    resultCodeActions.push(vscodeCodeAction);
-                                });
-                                return resultCodeActions;
-                            });
-                        }
-                    }
 
                     // Semantic token types are identified by indexes in this list of types, in the legend.
                     const tokenTypesLegend: string[] = [];
